+++ y3tu-tool-report/src/main/java/com/y3tu/tool/report/repository/DictSqlRepository.java
@@ -14,5 +14,5 @@ public interface DictSqlRepository extends BaseRepository<DictSql> {
      * @param dictId
      * @return
      */
-    DictSql findByDictId(long dictId);
+    DictSql findByDictId(int dictId);
 }
