+++ y3tu-tool-report/src/main/java/com/y3tu/tool/report/service/DictSqlService.java
@@ -13,5 +13,5 @@ public interface DictSqlService extends BaseService<DictSql> {
      * @param dictId
      * @return
      */
-    DictSql getByDictId(long dictId);
+    DictSql getByDictId(int dictId);
 }
